monitor rtm/thlm_spur_src at the CAM-CLUBB interface.
This version attempts to use mpi_reduce to get global max. conservation error
and the total number of columns in which a threshold is exceeded.
Only partically successful because tphysbc and tphysac operate only on a
single chunk (see physpkg.F90, subroutine phys_run1)

Nevertheless, results show that at nstep 0, the relative errors are
around 5e-16.